lolidk

+++ src/audio_core/time_stretch.h
@@ -19,6 +19,6 @@ void Shutdown();
 
 void Tick(unsigned samples_in_queue);
 void AddSamples(const std::array<std::array<s16, AudioCore::samples_per_frame>, 2>& samples);
-void OutputSamples(std::function<void(const std::vector<float>&)> fn);
+void OutputSamples(std::function<void(const std::vector<s16>&)> fn);
 
 }
